==> docs/html/files/notebooks/1.0-cbw-data-retrieval.ipynb <==
# Data retrieval from Kaggle API

import subprocess

# Works on Windows machines
def retrieve_data():
    """Remove contents from external data folder and download relevant Kaggle competition CSV files into it."""
    subprocess.run(['powershell', '-Command', 'Remove-Item ../../data/external/* -Recurse -Force'])
    subprocess.run('kaggle competitions download -c home-credit-default-risk -p ../../data/external/')
    subprocess.run(['powershell', '-Command',
                    'Expand-Archive ../../data/external/home-credit-default-risk.zip -DestinationPath ../../data/external/'])
    subprocess.run(['powershell', '-Command', 'rm ../../data/external/home-credit-default-risk.zip'])

retrieve_data()
